Move cerec_pose.* to mapping/pose_graph.
