+++ cartographer/mapping/pose_graph/ceres_pose.cc
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-#include "cartographer/mapping_3d/ceres_pose.h"
+#include "cartographer/mapping/pose_graph/ceres_pose.h"
 
 namespace cartographer {
-namespace mapping_3d {
+namespace mapping {
+namespace pose_graph {
 
 CeresPose::CeresPose(
     const transform::Rigid3d& rigid,
@@ -42,5 +43,6 @@ const transform::Rigid3d CeresPose::ToRigid() const {
       Eigen::Quaterniond(rotation[0], rotation[1], rotation[2], rotation[3]));
 }
 
-}  // namespace mapping_3d
+}  // namespace pose_graph
+}  // namespace mapping
 }  // namespace cartographer
