+++ cartographer/mapping/pose_graph/ceres_pose.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_MAPPING_3D_CERES_POSE_H_
-#define CARTOGRAPHER_MAPPING_3D_CERES_POSE_H_
+#ifndef CARTOGRAPHER_MAPPING_POSE_GRAPH_CERES_POSE_H_
+#define CARTOGRAPHER_MAPPING_POSE_GRAPH_CERES_POSE_H_
 
 #include <array>
 #include <memory>
@@ -25,7 +25,8 @@
 #include "ceres/ceres.h"
 
 namespace cartographer {
-namespace mapping_3d {
+namespace mapping {
+namespace pose_graph {
 
 class CeresPose {
  public:
@@ -49,7 +50,8 @@ class CeresPose {
   std::shared_ptr<Data> data_;
 };
 
-}  // namespace mapping_3d
+}  // namespace pose_graph
+}  // namespace mapping
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_MAPPING_3D_CERES_POSE_H_
+#endif  // CARTOGRAPHER_MAPPING_POSE_GRAPH_CERES_POSE_H_
